backend/requirements.txt
Django==5.1.6
gunicorn==23.0.0
psycopg2-binary==2.9.10
Pillow==11.1.0
pydantic==2.10.6
python-dotenv==1.0.1
whitenoise==6.9.0
openai==1.61.1
